requirements.txt
streamlit>=1.32.0
nltk>=3.8.0
scikit-learn>=1.3.0
pdfminer.six>=20221105
python-docx>=1.0.0
numpy>=1.24.0
